backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
groq==0.4.2
pydantic==2.6.1
httpx==0.27.0
python-dotenv==1.0.1
